requirements.txt
fastapi
uvicorn[standard]
motor
python-dotenv
PyJWT
python-multipart
